CLN: move pandas.computation -> pandas.core.computation
