+++ asv_bench/benchmarks/eval.py
@@ -1,6 +1,6 @@
 from .pandas_vb_common import *
 import pandas as pd
-import pandas.computation.expressions as expr
+import pandas.core.computation.expressions as expr
 
 
 class Eval(object):
