+++ pandas/core/computation/api.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 
-from pandas.computation.eval import eval
+from pandas.core.computation.eval import eval
 
 
 # deprecation, xref #13790
@@ -10,5 +10,5 @@ def Expr(*args, **kwargs):
     warnings.warn("pd.Expr is deprecated as it is not "
                   "applicable to user code",
                   FutureWarning, stacklevel=2)
-    from pandas.computation.expr import Expr
+    from pandas.core.computation.expr import Expr
     return Expr(*args, **kwargs)
