+++ pandas/core/computation/pytables.py
@@ -10,10 +10,10 @@ import pandas.core.common as com
 from pandas.compat import u, string_types, DeepChainMap
 from pandas.core.base import StringMixin
 from pandas.formats.printing import pprint_thing, pprint_thing_encoded
-from pandas.computation import expr, ops
-from pandas.computation.ops import is_term, UndefinedVariableError
-from pandas.computation.expr import BaseExprVisitor
-from pandas.computation.common import _ensure_decoded
+from pandas.core.computation import expr, ops
+from pandas.core.computation.ops import is_term, UndefinedVariableError
+from pandas.core.computation.expr import BaseExprVisitor
+from pandas.core.computation.common import _ensure_decoded
 from pandas.tseries.timedeltas import _coerce_scalar_to_timedelta_type
