+++ pandas/io/pytables.py
@@ -44,7 +44,7 @@ from pandas.core.index import _ensure_index
 from pandas import compat
 from pandas.compat import u_safe as u, PY3, range, lrange, string_types, filter
 from pandas.core.config import get_option
-from pandas.computation.pytables import Expr, maybe_expression
+from pandas.core.computation.pytables import Expr, maybe_expression
 
 from pandas._libs import tslib, algos, lib
