+++ pandas/tests/test_expressions.py
@@ -12,7 +12,7 @@ from numpy.random import randn
 import numpy as np
 
 from pandas.core.api import DataFrame, Panel
-from pandas.computation import expressions as expr
+from pandas.core.computation import expressions as expr
 from pandas import compat, _np_version_under1p11
 from pandas.util.testing import (assert_almost_equal, assert_series_equal,
                                  assert_frame_equal, assert_panel_equal,
