+++ setup.py
@@ -639,9 +639,9 @@ setup(name=DISTNAME,
                 'pandas.api.lib',
                 'pandas.compat',
                 'pandas.compat.numpy',
-                'pandas.computation',
                 'pandas.core',
                 'pandas.core.typed',
+                'pandas.core.computation',
                 'pandas.indexes',
                 'pandas.errors',
                 'pandas.io',
@@ -656,8 +656,8 @@ setup(name=DISTNAME,
                 'pandas.util',
                 'pandas.tests',
                 'pandas.tests.api',
-                'pandas.tests.computation',
                 'pandas.tests.core.typed',
+                'pandas.tests.core.computation',
                 'pandas.tests.frame',
                 'pandas.tests.indexes',
                 'pandas.tests.indexes.datetimes',
